requirements.txt
pyyaml>=6.0
Jinja2>=3.0.0
click>=8.0.0
python-dateutil>=2.8.0
tabulate>=0.9.0
colorama>=0.4.4
